test(Operators): add tests for bad arguments, split tests in separate files
