+++ packages/cerebral/src/operators/unshift.test.js
@@ -1,0 +1,49 @@
+/* eslint-env mocha */
+import Controller from '../Controller'
+import assert from 'assert'
+import {input, state, unshift} from './'
+
+describe('operators.unshift', () => {
+  it('should unshift value in model', () => {
+    const controller = new Controller({
+      state: {
+        list: ['a', 'b']
+      },
+      signals: {
+        test: [
+          unshift(state`list`, 'x')
+        ]
+      }
+    })
+    controller.getSignal('test')()
+    assert.deepEqual(controller.getState(), {list: ['x', 'a', 'b']})
+  })
+  it('should unshift value from input in model', () => {
+    const controller = new Controller({
+      state: {
+        list: ['a', 'b']
+      },
+      signals: {
+        test: [
+          unshift(state`list`, input`value`)
+        ]
+      }
+    })
+    controller.getSignal('test')({value: 'x'})
+    assert.deepEqual(controller.getState(), {list: ['x', 'a', 'b']})
+  })
+  it('should throw on bad argument', () => {
+    const controller = new Controller({
+      state: {
+      },
+      signals: {
+        test: [
+          unshift(input`list`, 'bar')
+        ]
+      }
+    })
+    assert.throws(() => {
+      controller.getSignal('test')({list: ['one']})
+    }, /operator.unshift/)
+  })
+})
